+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Basic support for DRM on iOS and Android [#1445](https://github.com/react-native-community/react-native-video/pull/1445)
 * Upgrade ExoPlayer to from 2.9.3 to 2.10.3
 * Added media session for Android ExoPlayer allowing user to pause/play playback with headset HW buttons, or for example display player controls in PIP window. [#1780](https://github.com/react-native-community/react-native-video/pull/1780)
+* Added PictureInPicture support for AndroidExoplayer API >= 28. [#1776](https://github.com/react-native-community/react-native-video/pull/1776)
 
 ### Version 5.0.2
 * Fix crash when RCTVideo's superclass doesn't observe the keyPath 'frame' (iOS) [#1720](https://github.com/react-native-community/react-native-video/pull/1720)